refactor: IPNE 描画の戦闘エフェクト処理を combatEffects へ分離
- 攻撃/被弾トリガー・外部キュー・エフェクト更新描画・floating text を逐語移植
- 副作用（ref 変更）と呼び出し位置（敵とプレイヤーの間）を厳守
- 特性化テストで描画コマンド列の不変を確認
